Apply Supabase best practices: batch operations, bigint PK, restrict refresh
- Increase batch size from 500 to 1000 rows across all sync scripts
- Batch since-assessment updates in sync-gbif (was row-by-row, now upsert)
- Batch deletes in sync-redlist deleteDelisted (was row-by-row, now .in())
- Restrict refresh_taxa_summary() to service_role only
- Use bigint for species.id per Supabase PK best practices

+++ supabase/migrations/20260306084638_species_link.sql
@@ -1,6 +1,6 @@
 -- Linking/identity table (joins IUCN and GBIF data)
 create table public.species (
-  id               integer generated always as identity primary key,
+  id               bigint generated always as identity primary key,
   sis_taxon_id     integer unique references public.redlist_species(sis_taxon_id) on delete set null,
   gbif_species_key integer unique references public.gbif_species(gbif_species_key) on delete set null
 );
